separate out thread loading from get state
