+++ hive/server/bridge_api/thread.py
@@ -1,0 +1,86 @@
+"""Routes then builds a get_state response object"""
+
+import logging
+
+from hive.server.common.mutes import Mutes
+from hive.server.bridge_api.objects import load_posts_keyed
+from hive.server.common.helpers import (
+    return_error_info,
+    valid_account,
+    valid_permlink)
+
+log = logging.getLogger(__name__)
+
+@return_error_info
+async def get_discussion(context, author, permlink):
+    """Modified `get_state` thread implementation."""
+    db = context['db']
+
+    author = valid_account(author)
+    permlink = valid_permlink(permlink)
+    root_id = await _get_post_id(db, author, permlink)
+    if not root_id:
+        return {}
+
+    return await _load_discussion(db, root_id)
+
+async def _get_post_id(db, author, permlink):
+    """Given an author/permlink, retrieve the id from db."""
+    sql = ("SELECT id FROM hive_posts WHERE author = :a "
+           "AND permlink = :p AND is_deleted = '0' LIMIT 1")
+    return await db.query_one(sql, a=author, p=permlink)
+
+def _ref(post):
+    return post['author'] + '/' + post['permlink']
+
+async def _child_ids(db, parent_ids):
+    """Load child ids for multuple parent ids."""
+    sql = """
+             SELECT parent_id, array_agg(id)
+               FROM hive_posts
+              WHERE parent_id IN :ids
+                AND is_deleted = '0'
+           GROUP BY parent_id
+    """
+    rows = await db.query_all(sql, ids=tuple(parent_ids))
+    return [[row[0], row[1]] for row in rows]
+
+async def _load_discussion(db, root_id):
+    """Load a full discussion thread."""
+    # build `ids` list and `tree` map
+    ids = []
+    tree = {}
+    todo = [root_id]
+    while todo:
+        ids.extend(todo)
+        rows = await _child_ids(db, todo)
+        todo = []
+        for pid, cids in rows:
+            tree[pid] = cids
+            todo.extend(cids)
+
+    # load all post objects, build ref-map
+    posts = await load_posts_keyed(db, ids)
+
+    # remove posts/comments from muted accounts
+    muted_accounts = Mutes.all()
+    rem_pids = []
+    for pid, post in posts.items():
+        if post['author'] in muted_accounts:
+            rem_pids.append(pid)
+    for pid in rem_pids:
+        if pid in posts:
+            del posts[pid]
+        if pid in tree:
+            rem_pids.extend(tree[pid])
+
+    refs = {pid: _ref(post) for pid, post in posts.items()}
+
+    # add child refs to parent posts
+    for pid, post in posts.items():
+        if pid in tree:
+            post['replies'] = [refs[cid] for cid in tree[pid]
+                               if cid in refs]
+
+    # return all nodes keyed by ref
+    return {refs[pid]: post for pid, post in posts.items()}
